Issue #58 Naming of repeat methods; additional tests; fix hard-coded GBP.
