+++ tests/Message/DirectRepeatAuthorizeRequestTest.php
@@ -4,10 +4,10 @@ namespace Omnipay\SagePay\Message;
 
 use Omnipay\Tests\TestCase;
 
-class DirectRepeatPaymentRequestTest extends TestCase
+class DirectRepeatAuthorizeRequestTest extends TestCase
 {
     /**
-     * @var \Omnipay\SagePay\Message\DirectRepeatPaymentRequest $request
+     * @var \Omnipay\SagePay\Message\DirectRepeatAuthorizeRequest $request
      */
     protected $request;
 
@@ -15,7 +15,7 @@ class DirectRepeatPaymentRequestTest extends TestCase
     {
         parent::setUp();
 
-        $this->request = new DirectRepeatPaymentRequest($this->getHttpClient(), $this->getHttpRequest());
+        $this->request = new DirectRepeatAuthorizeRequest($this->getHttpClient(), $this->getHttpRequest());
         $this->request->initialize(
             array(
                 'amount' => '12.00',
