+++ tests/Mock/RepeatAuthorize.txt
@@ -1,0 +1,22 @@
+HTTP/1.1 200 OK
+Date: Sat, 16 Feb 2013 06:39:41 GMT
+Server: Microsoft-IIS/6.0
+P3P: CP="CUR"
+X-Powered-By: ASP.NET
+Content-Length: 249
+Content-Type: Text/Plain
+Set-Cookie: ASPSESSIONIDSEHSCTTR=CIMGGPBDFDIKICAGIIHCNFHJ; secure; path=/
+Cache-control: private
+
+VPSProtocol=3.00
+Status=OK
+StatusDetail=Successful repeat.
+VPSTxId={5A1BC414-5409-48DD-9B8B-DCDF096CE0BE}
+SecurityKey=OUWLNYQTVT
+TxAuthNo=9962
+AVSCV2=ALL MATCH
+AddressResult=MATCHED
+PostCodeResult=MATCHED
+CV2Result=MATCHED
+BankAuthCode=T99777
+DeclineCode=00
